Refactoring. Separate components from core. Replaces enum ComponentBit with and int so the core doesn't have to know anything about component flags.
Rename system.* and component.* to systems.* and components.* because that's what they are. Pural.

Introduce a component base in core which houses the Component struct, defining its flag, alignment and size because that part of components is general and the entity component system uses this to be able to make component pools.